remove options components for better control of play interface
