+++ lib/components/index.js
@@ -2,7 +2,6 @@ import MainScreenHeader from './MainScreenHeader';
 import Browse from './Browse';
 import PlaylistCard from './PlaylistCard';
 import Search from './Search';
-import Options from './Options';
 import BackButton from './BackButton';
 import OptionButton from './OptionButton';
 import GameStatusBoard from './GameStatusBoard';
@@ -13,7 +12,6 @@ export {
   Browse,
   PlaylistCard,
   Search,
-  Options,
   BackButton,
   OptionButton,
   GameStatusBoard,
